flurry-back/requirements.txt
beautifulsoup4==4.13.4
google-genai==1.11.0
google-generativeai==0.8.5
groq==0.22.0
protobuf==5.29.4
pydub==0.25.1
python-dotenv==1.1.0
requests==2.32.3
tiktoken==0.9.0